Predictive model construction using R codes
Paper title: Predicting the Reproductive Toxicity of Chemicals Using Ensemble Learning Methods and Molecular Fingerprints
Abstract: Reproductive toxicity endpoints are a significant safety concern in the assessment of adverse effects of chemicals in drug discovery. Computational models that can accurately predict a chemical's toxic potential on computers are increasingly pursued to replace traditional animal experiments. To this end, ensemble learning models were built to predict the reproductive toxicity of compounds. Our ensemble models were developed using Support vector machine, Random forest, and Extreme gradient boosting and 9 molecular fingerprints calculated for a dataset containing 1823 chemicals. The best prediction performance was achieved by the Ensemble-Top12 model, with an accuracy (ACC) of 86.33%, a sensitivity (SEN) of 82.02%, a specificity (SPE) of 90.19%, and an area under the receiver operating characteristic curve (AUC) of 0.937 in 5-fold cross-validation and an ACC of 84.38%, a SEN of 86.90%, a SPE of 90.67%, and an AUC of 0.920 in an external validation. This ensemble model achieves relatively high ACC, SPE and AUC values compared with methods previously reported in the literature. Furthermore, several fingerprint features related to chemical reproductive toxicity are identified. Overall, our research has effectively improved the capacity of computational models to predict the reproductive toxicity of chemicals and will be helpful for the early evaluation of chemical reproductive toxicity in drug-hazard assessments.